chore: sync ci scaffolding from duck-mc + relax clippy
- LICENSE, CONTRIBUTING, CODE_OF_CONDUCT, SECURITY
- .editorconfig, .typos.toml, deny.toml, release-plz.toml
- .github/workflows/{ci,audit,codeql,typos,scorecard,release}.yml
- .github/{labels.sh,protect-branches.sh,PULL_REQUEST_TEMPLATE.md,
            dependabot.yml,ISSUE_TEMPLATE/}
- .husky/pre-push: mirror CI gates locally
- README.md: rewrite in duck-mc style
- Cargo.toml workspace.lints.clippy: allow noisy style lints (existing
  code has tech debt; tightening is a follow-up).
- ci.yml + pre-push run clippy without -D warnings until cleanup PR
  lands.

+++ ttlog-view/src/logs_chart_widget/mod.rs
@@ -207,7 +207,7 @@ impl<'a> LogsChartWidget<'a> {
   /// and bottom shortcuts in right panel with separator + padding.
   fn render_colored_bars(&mut self, f: &mut Frame<'_>, area: Rect) {
     let bars = self.build_bars();
-    let n = bars.len() as usize;
+    let n = bars.len();
     let (total, _, _, _, _, _, _) = self.aggregate_counts();
     let max_value = bars.iter().map(|(_, v, _)| *v).max().unwrap_or(1).max(1);
